# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Microcredit management system API for Argentina",
        "title": "Crédito Villanueva API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@creditovillanueva.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/admin/credit-lines": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new credit line for a client with specified limits and interest rate",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Credit Lines"
                ],
                "summary": "Create a credit line",
                "parameters": [
                    {
                        "description": "Credit line data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateCreditLineRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.CreditLineResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/credit-lines/pending": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a paginated list of credit lines awaiting approval",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Credit Lines"
                ],
                "summary": "List pending credit lines",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Pagination offset",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Pagination limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.PaginatedResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/credit-lines/{id}/approve": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Approves a pending credit line, making it available for the client to request loans",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Credit Lines"
                ],
                "summary": "Approve a credit line",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Credit Line UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.CreditLineResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/credit-lines/{id}/reject": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Rejects a pending credit line with a reason",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Credit Lines"
                ],
                "summary": "Reject a credit line",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Credit Line UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Rejection reason",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RejectCreditLineRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.CreditLineResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/dashboard/delinquency": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns delinquency metrics including PAR30, PAR60, PAR90, and overall delinquency rate",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get delinquency statistics",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.DelinquencyResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/dashboard/kpis": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns combined portfolio and delinquency KPIs in a single response",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get combined KPIs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.KPIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/dashboard/portfolio": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns portfolio statistics including total clients, active loans, disbursed and outstanding amounts",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get portfolio summary",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.PortfolioResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/dashboard/trends/collections": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns monthly collection trend data for the last 12 months",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get collection trends",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.TrendPointResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/dashboard/trends/disbursements": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns monthly disbursement trend data for the last 12 months",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get disbursement trends",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.TrendPointResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CreateCreditLineRequest": {
            "type": "object",
            "required": [
                "clientId",
                "interestRate",
                "maxAmount",
                "maxInstallments"
            ],
            "properties": {
                "clientId": {
                    "type": "string"
                },
                "interestRate": {
                    "type": "string"
                },
                "maxAmount": {
                    "type": "string"
                },
                "maxInstallments": {
                    "type": "integer",
                    "maximum": 60,
                    "minimum": 1
                }
            }
        },
        "dto.CreditLineResponse": {
            "type": "object",
            "properties": {
                "approvedAt": {
                    "type": "string"
                },
                "availableAmount": {
                    "type": "string"
                },
                "clientId": {
                    "type": "string"
                },
                "clientName": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "interestRate": {
                    "type": "string"
                },
                "maxAmount": {
                    "type": "string"
                },
                "maxInstallments": {
                    "type": "integer"
                },
                "rejectionReason": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "usedAmount": {
                    "type": "string"
                }
            }
        },
        "dto.DelinquencyResponse": {
            "type": "object",
            "properties": {
                "delinquencyRate": {
                    "type": "string"
                },
                "overdueCount": {
                    "type": "integer"
                },
                "par30": {
                    "type": "string"
                },
                "par60": {
                    "type": "string"
                },
                "par90": {
                    "type": "string"
                },
                "totalOverdue": {
                    "type": "string"
                }
            }
        },
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "dto.KPIResponse": {
            "type": "object",
            "properties": {
                "delinquency": {
                    "$ref": "#/definitions/dto.DelinquencyResponse"
                },
                "portfolio": {
                    "$ref": "#/definitions/dto.PortfolioResponse"
                }
            }
        },
...
        "dto.PaginatedResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "dto.PortfolioResponse": {
            "type": "object",
            "properties": {
                "activeLoans": {
                    "type": "integer"
                },
                "pendingApprovals": {
                    "type": "integer"
                },
                "totalClients": {
                    "type": "integer"
                },
                "totalCollected": {
                    "type": "string"
                },
                "totalDisbursed": {
                    "type": "string"
                },
                "totalOutstanding": {
                    "type": "string"
                }
            }
        },
...
        "dto.RejectCreditLineRequest": {
            "type": "object",
            "required": [
                "reason"
            ],
            "properties": {
                "reason": {
                    "type": "string"
                }
            }
        },
...
        "dto.TrendPointResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string"
                },
                "count": {
                    "type": "integer"
                },
                "date": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter your bearer token in the format: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}